+++ jsonpath/env.py
@@ -70,6 +70,8 @@ class JSONPathEnvironment:
     ## Class attributes
 
     Attributes:
+        filter_caching (bool): If `True`, filter expressions will be cached where
+            possible.
         filter_context_token (str): The pattern used to select extra filter context
             data. Defaults to `"_"`.
         intersection_token (str): The pattern used as the intersection operator.
@@ -104,12 +106,14 @@ class JSONPathEnvironment:
     max_int_index = (2**53) - 1
     min_int_index = -(2**53) + 1
 
+    filter_caching = False
+
     # Override these to customize path tokenization and parsing.
     lexer_class: Type[Lexer] = Lexer
     parser_class: Type[Parser] = Parser
     match_class: Type[JSONPathMatch] = JSONPathMatch
 
-    def __init__(self) -> None:  # noqa: D107
+    def __init__(self) -> None:
         self.lexer: Lexer = self.lexer_class(env=self)
         """The lexer bound to this environment."""